Implemented fuel regen shipstat (and changed fuel stuff to doubles again...)

+++ src/pilot.h
@@ -243,9 +243,9 @@ typedef struct Pilot_ {
    double armour_regen; /**< Armour regeneration rate (per second). */
    double shield_regen; /**< Shield regeneration rate (per second). */
    double dmg_absorb; /**< Ship damage absorption [0:1] with 1 being 100%. */
-   int    fuel_max;   /**< Maximum fuel. */
-   int    fuel;       /**< Current fuel. */
-   int    fuel_consumption; /**< Fuel consumed per jump. */
+   double fuel_max;   /**< Maximum fuel. */
+   double fuel;       /**< Current fuel. */
+   double fuel_consumption; /**< Fuel consumed per jump. */
 
    /* Energy is handled a bit differently. */
    double energy;    /**< Current energy. */
